+++ gui/src/redux/thunks/streamResponse.test.ts
@@ -1,102 +1,82 @@
 import { describe, expect, it, vi } from "vitest";
-import { JSONContent } from "@tiptap/core";
-import { InputModifiers } from "core";
 import { createMockStore } from "../../util/test/mockStore";
-import { streamResponseThunk } from "./streamResponse";
+import { streamNormalInput } from "./streamNormalInput";
 
-// Mock only the external dependencies - let thunks run naturally
-vi.mock("posthog-js", () => ({
-  default: {
-    capture: vi.fn(),
+// Mock external dependencies only - let selectors run naturally
+vi.mock("core/llm/toolSupport", () => ({
+  modelSupportsNativeTools: vi.fn(),
+}));
+
+vi.mock(
+  "core/tools/systemMessageTools/buildToolsSystemMessage",
+  async (importOriginal) => {
+    const actual = (await importOriginal()) as any;
+    return {
+      ...actual,
+      addSystemMessageToolsToSystemMessage: vi.fn(),
+    };
   },
-}));
-
-vi.mock("uuid", () => ({
-  v4: vi.fn(() => "mock-uuid-123"),
-}));
-
-vi.mock("../../components/mainInput/TipTapEditor/utils/resolveEditorContent", () => ({
-  resolveEditorContent: vi.fn(),
-}));
-
-// Mock selectors that streamNormalInput uses
-vi.mock("../selectors/selectActiveTools", () => ({
-  selectActiveTools: vi.fn(() => []),
-}));
-
-vi.mock("../selectors/selectUseSystemMessageTools", () => ({
-  selectUseSystemMessageTools: vi.fn(() => false),
-}));
+);
 
 vi.mock("../util/constructMessages", () => ({
-  constructMessages: vi.fn(() => ({
-    messages: [{ role: "user", content: "Hello" }],
-    appliedRules: [],
-    appliedRuleIndex: 0,
-  })),
+  constructMessages: vi.fn(),
 }));
 
 vi.mock("../util/getBaseSystemMessage", () => ({
-  getBaseSystemMessage: vi.fn(() => "You are a helpful assistant"),
-}));
-
-vi.mock("core/llm/toolSupport", () => ({
-  modelSupportsNativeTools: vi.fn(() => true),
+  getBaseSystemMessage: vi.fn(),
 }));
 
 vi.mock("core/config/shouldAutoEnableSystemMessageTools", () => ({
   shouldAutoEnableSystemMessageTools: vi.fn(() => undefined),
 }));
 
-import posthog from "posthog-js";
-import { resolveEditorContent } from "../../components/mainInput/TipTapEditor/utils/resolveEditorContent";
 import { ModelDescription } from "core";
+import { modelSupportsNativeTools } from "core/llm/toolSupport";
+import { addSystemMessageToolsToSystemMessage } from "core/tools/systemMessageTools/buildToolsSystemMessage";
+import { constructMessages } from "../util/constructMessages";
+import { getBaseSystemMessage } from "../util/getBaseSystemMessage";
 
-const mockPosthog = vi.mocked(posthog);
-const mockResolveEditorContent = vi.mocked(resolveEditorContent);
+const mockModelSupportsNativeTools = vi.mocked(modelSupportsNativeTools);
+const mockAddSystemMessageToolsToSystemMessage = vi.mocked(
+  addSystemMessageToolsToSystemMessage,
+);
+const mockConstructMessages = vi.mocked(constructMessages);
+const mockGetBaseSystemMessage = vi.mocked(getBaseSystemMessage);
 
 const mockClaudeModel: ModelDescription = {
   title: "Claude 3.5 Sonnet",
   model: "claude-3-5-sonnet-20241022",
   provider: "anthropic",
   underlyingProviderName: "anthropic",
-  completionOptions: {
-    reasoningBudgetTokens: 2048,
-  },
-};
-
-const mockEditorState: JSONContent = {
-  type: "doc",
-  content: [
-    {
-      type: "paragraph",
-      content: [{ type: "text", text: "Hello, please help me with this code" }],
-    },
-  ],
-};
-
-const mockModifiers: InputModifiers = {
-  useCodebase: true,
-  noContext: false,
+  completionOptions: { reasoningBudgetTokens: 2048 },
 };
 
 function setupTest() {
   vi.clearAllMocks();
 
+  // Default mock implementations for external functions
+  mockModelSupportsNativeTools.mockReturnValue(true);
+  mockGetBaseSystemMessage.mockReturnValue("System message");
+  mockAddSystemMessageToolsToSystemMessage.mockReturnValue(
+    "System message with tools",
+  );
+  mockConstructMessages.mockReturnValue({
+    messages: [{ role: "user", content: "Hello" }],
+    appliedRules: [],
+    appliedRuleIndex: 0,
+  });
+
+  // Create store with realistic state that selectors can work with
   const mockStore = createMockStore({
     session: {
       history: [
         {
-          message: {
-            id: "1",
-            role: "user",
-            content: "Previous message",
-          },
+          message: { id: "1", role: "user", content: "Hello" },
           contextItems: [],
         },
       ],
       hasReasoningEnabled: false,
-      isStreaming: false,
+      isStreaming: true,
       id: "session-123",
       mode: "chat",
       streamAborter: new AbortController(),
@@ -120,7 +100,6 @@ function setupTest() {
       config: {
         tools: [],
         rules: [],
-        slashCommands: [],
         tabAutocompleteModel: undefined,
         selectedModelByRole: {
           chat: mockClaudeModel,
@@ -132,7 +111,6 @@ function setupTest() {
           embed: null,
         },
         experimental: {
-          defaultContext: ["codebase"],
           onlyUseSystemMessageTools: false,
         },
       } as any,
@@ -148,59 +126,2455 @@ function setupTest() {
     } as any,
   });
 
-  // Setup IDE messenger responses for streamNormalInput
-  mockStore.mockIdeMessenger.request.mockImplementation((endpoint: string) => {
-    if (endpoint === "llm/compileChat") {
-      return Promise.resolve({
-        status: "success",
-        content: {
-          compiledChatMessages: [{ role: "user", content: "Hello" }],
-          didPrune: false,
-          contextPercentage: 0.8,
-        },
-      });
-    } else if (endpoint === "history/save") {
-      return Promise.resolve({ status: "success" });
-    }
-    return Promise.resolve({ status: "success", content: {} });
-  });
+  const mockIdeMessenger = mockStore.mockIdeMessenger;
 
-  // Setup streaming generator for streamNormalInput
-  async function* mockStreamGenerator() {
-    yield [{ role: "assistant", content: "I'll help you with that." }];
-    return {
-      prompt: "Hello",
-      completion: "I'll help you with that.",
-      modelProvider: "anthropic",
-    };
-  }
-
-  mockStore.mockIdeMessenger.llmStreamChat.mockReturnValue(mockStreamGenerator());
-
-  return { mockStore };
+  return { mockStore, mockIdeMessenger };
 }
 
-describe("streamResponseThunk", () => {
-  it("should execute complete chat flow without slash commands", async () => {
-    const { mockStore } = setupTest();
+describe("streamNormalInput", () => {
+  it("should execute complete streaming flow with all dispatches", async () => {
+    const { mockStore, mockIdeMessenger } = setupTest();
+    // Setup successful compilation
+    mockIdeMessenger.request.mockResolvedValue({
+      status: "success",
+      content: {
+        compiledChatMessages: [{ role: "user", content: "Hello" }],
+        didPrune: false,
+        contextPercentage: 0.8,
+      },
+    });
 
-    // Mock resolveEditorContent for basic chat
-    mockResolveEditorContent.mockResolvedValue({
-      selectedContextItems: [],
-      selectedCode: [],
-      content: "Hello, please help me with this code",
-      legacyCommandWithInput: null,
+    // Setup streaming generator
+    async function* mockStreamGenerator() {
+      yield [{ role: "assistant", content: "First chunk" }];
+      yield [{ role: "assistant", content: "Second chunk" }];
+      return {
+        prompt: "Hello",
+        completion: "Hi there!",
+        modelProvider: "anthropic",
+      };
+    }
+
+    mockIdeMessenger.llmStreamChat.mockReturnValue(mockStreamGenerator());
+
+    // Execute thunk
+    const result = await mockStore.dispatch(streamNormalInput({}) as any);
+
+    // Verify exact sequence of dispatched actions with payloads
+    const dispatchedActions = (mockStore as any).getActions();
+
+    expect(dispatchedActions).toEqual([
+      {
+        type: "chat/streamNormalInput/pending",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "pending",
+        }),
+      },
+      {
+        type: "session/setAppliedRulesAtIndex",
+        payload: {
+          index: 0,
+          appliedRules: [],
+        },
+      },
+      {
+        type: "session/setActive",
+      },
+      {
+        type: "session/setInlineErrorMessage",
+      },
+      {
+        type: "session/setIsPruned",
+        payload: false,
+      },
+      {
+        type: "session/setContextPercentage",
+        payload: 0.8,
+      },
+      {
+        type: "session/streamUpdate",
+        payload: [
+          {
+            role: "assistant",
+            content: "First chunk",
+          },
+        ],
+      },
+      {
+        type: "session/streamUpdate",
+        payload: [
+          {
+            role: "assistant",
+            content: "Second chunk",
+          },
+        ],
+      },
+      {
+        type: "session/addPromptCompletionPair",
+        payload: [
+          {
+            prompt: "Hello",
+            completion: "Hi there!",
+            modelProvider: "anthropic",
+          },
+        ],
+      },
+      {
+        type: "session/setInactive",
+      },
+      {
+        type: "chat/streamNormalInput/fulfilled",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "fulfilled",
+        }),
+      },
+    ]);
+
+    // Verify IDE messenger calls
+    expect(mockIdeMessenger.request).toHaveBeenCalledWith("llm/compileChat", {
+      messages: [{ role: "user", content: "Hello" }],
+      options: {},
+    });
+
+    expect(mockIdeMessenger.llmStreamChat).toHaveBeenCalledWith(
+      expect.objectContaining({
+        completionOptions: {},
+        title: "Claude 3.5 Sonnet",
+        messages: [{ role: "user", content: "Hello" }],
+        messageOptions: { precompiled: true },
+      }),
+      expect.any(AbortSignal),
+    );
+
+    // Verify dev data logging call
+    expect(mockIdeMessenger.post).toHaveBeenCalledWith("devdata/log", {
+      name: "chatInteraction",
+      data: {
+        prompt: "Hello",
+        completion: "Hi there!",
+        modelProvider: "anthropic",
+        modelTitle: "Claude 3.5 Sonnet",
+        sessionId: "session-123",
+      },
+    });
+
+    expect(result.type).toBe("chat/streamNormalInput/fulfilled");
+
+    // Verify final state after thunk completion
+    const finalState = mockStore.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            appliedRules: [],
+            contextItems: [],
+            message: { id: "1", role: "user", content: "Hello" },
+          },
+          {
+            contextItems: [],
+            isGatheringContext: false,
+            message: {
+              content: "First chunkSecond chunk", // Chunks get combined
+              id: expect.any(String),
+              role: "assistant",
+            },
+            promptLogs: [
+              {
+                completion: "Hi there!",
+                modelProvider: "anthropic",
+                prompt: "Hello",
+              },
+            ],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: false,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController),
+        contextPercentage: 0.8,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {},
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
+  });
+
+  it("should execute streaming flow with tool call execution", async () => {
+    const { mockStore, mockIdeMessenger } = setupTest();
+
+    // Set up auto-approved tool setting for our test tool
+    const mockStoreWithToolSettings = createMockStore({
+      session: {
+        history: [
+          {
+            message: {
+              id: "1",
+              role: "user",
+              content: "Please search the codebase",
+            },
+            contextItems: [],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: true,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: new AbortController(),
+        contextPercentage: 0,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        } as any,
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+      } as any,
+      ui: {
+        toolSettings: {
+          search_codebase: "allowedWithoutPermission", // Auto-approve this tool
+        },
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      } as any,
+    });
+
+    const mockIdeMessengerWithTool = mockStoreWithToolSettings.mockIdeMessenger;
+
+    // Setup successful compilation and tool responses
+    mockIdeMessengerWithTool.request.mockImplementation(
+      (endpoint: string, data: any) => {
+        if (endpoint === "llm/compileChat") {
+          return Promise.resolve({
+            status: "success",
+            content: {
+              compiledChatMessages: [
+                { role: "user", content: "Please search the codebase" },
+              ],
+              didPrune: false,
+              contextPercentage: 0.9,
+            },
+          });
+        } else if (endpoint === "tools/call") {
+          // Mock server-side tool call response
+          return Promise.resolve({
+            status: "success",
+            content: {
+              contextItems: [
+                {
+                  name: "Search Results",
+                  description: "Found 3 matches",
+                  content: "Result 1\nResult 2\nResult 3",
+                  icon: "search",
+                  hidden: false,
+                },
+              ],
+              errorMessage: undefined,
+            },
+          });
+        } else if (endpoint === "history/save") {
+          // Mock session save
+          return Promise.resolve({ status: "success" });
+        }
+        // Let other endpoints pass through or return success
+        return Promise.resolve({ status: "success", content: {} });
+      },
+    );
+
+    // Setup streaming generator with tool call
+    async function* mockStreamGeneratorWithTool() {
+      yield [
+        { role: "assistant", content: "I'll search the codebase for you." },
+      ];
+      yield [
+        {
+          role: "assistant",
+          content: "",
+          toolCalls: [
+            {
+              id: "tool-call-1",
+              type: "function",
+              function: {
+                name: "search_codebase",
+                arguments: JSON.stringify({ query: "test function" }),
+              },
+            },
+          ],
+        },
+      ];
+      return {
+        prompt: "Please search the codebase",
+        completion: "I'll search the codebase for you.",
+        modelProvider: "anthropic",
+      };
+    }
+
+    // Mock different streaming responses for multiple calls
+    let streamCallCount = 0;
+    mockIdeMessengerWithTool.llmStreamChat.mockImplementation(() => {
+      streamCallCount++;
+      if (streamCallCount === 1) {
+        // First call - main streaming with tool call
+        return mockStreamGeneratorWithTool();
+      } else {
+        // Subsequent calls from streamResponseAfterToolCall - return minimal response
+        async function* simpleGenerator() {
+          yield [{ role: "assistant", content: "Search completed." }];
+          return {
+            prompt: "continuing after tool",
+            completion: "Search completed.",
+            modelProvider: "anthropic",
+          };
+        }
+        return simpleGenerator();
+      }
     });
 
     // Execute thunk
-    const result = await mockStore.dispatch(
-      streamResponseThunk({
-        editorState: mockEditorState,
-        modifiers: mockModifiers,
-      }) as any,
+    const result = await mockStoreWithToolSettings.dispatch(
+      streamNormalInput({}) as any,
     );
 
-    // Verify thunk completed successfully
-    expect(result.type).toBe("chat/streamResponse/fulfilled");
+    // Verify key actions are dispatched (tool calls trigger a complex cascade, so we verify key actions exist)
+    const dispatchedActions = (mockStoreWithToolSettings as any).getActions();
+
+    // Verify core streaming actions
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setAppliedRulesAtIndex",
+        payload: { index: 0, appliedRules: [] },
+      }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({ type: "session/setActive" }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setContextPercentage",
+        payload: 0.9,
+      }),
+    );
+
+    // Verify streaming updates (both regular content and tool call)
+    const streamUpdates = dispatchedActions.filter(
+      (action: any) => action.type === "session/streamUpdate",
+    );
+    expect(streamUpdates.length).toBeGreaterThanOrEqual(2);
+    expect(streamUpdates[0].payload).toEqual([
+      { role: "assistant", content: "I'll search the codebase for you." },
+    ]);
+    expect(streamUpdates[1].payload[0]).toEqual(
+      expect.objectContaining({
+        role: "assistant",
+        content: "",
+        toolCalls: [
+          {
+            id: "tool-call-1",
+            type: "function",
+            function: {
+              name: "search_codebase",
+              arguments: JSON.stringify({ query: "test function" }),
+            },
+          },
+        ],
+      }),
+    );
+
+    // Verify tool call execution actions
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setToolGenerated",
+        payload: expect.objectContaining({ toolCallId: "tool-call-1" }),
+      }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setToolCallCalling",
+        payload: { toolCallId: "tool-call-1" },
+      }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/updateToolCallOutput",
+        payload: {
+          toolCallId: "tool-call-1",
+          contextItems: [
+            {
+              name: "Search Results",
+              description: "Found 3 matches",
+              content: "Result 1\nResult 2\nResult 3",
+              icon: "search",
+              hidden: false,
+            },
+          ],
+        },
+      }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/acceptToolCall",
+        payload: { toolCallId: "tool-call-1" },
+      }),
+    );
+
+    // Verify thunk completion
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "chat/streamNormalInput/fulfilled",
+      }),
+    );
+
+    // Verify IDE messenger calls
+    expect(mockIdeMessengerWithTool.request).toHaveBeenCalledWith(
+      "llm/compileChat",
+      {
+        messages: [{ role: "user", content: "Hello" }], // constructMessages mock returns this
+        options: {},
+      },
+    );
+
+    expect(mockIdeMessengerWithTool.request).toHaveBeenCalledWith(
+      "tools/call",
+      {
+        toolCall: {
+          id: "tool-call-1",
+          type: "function",
+          function: {
+            name: "search_codebase",
+            arguments: JSON.stringify({ query: "test function" }),
+          },
+        },
+      },
+    );
+
+    // Verify that multiple compilation calls were made (due to tool call continuation)
+    const compileCallsCount =
+      mockIdeMessengerWithTool.request.mock.calls.filter(
+        (call: any) => call[0] === "llm/compileChat",
+      ).length;
+    expect(compileCallsCount).toBeGreaterThanOrEqual(1);
+
+    expect(result.type).toBe("chat/streamNormalInput/fulfilled");
+
+    // Verify final state after tool call execution
+    const finalState = mockStoreWithToolSettings.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            appliedRules: [],
+            contextItems: [],
+            message: {
+              id: "1",
+              role: "user",
+              content: "Please search the codebase",
+            },
+          },
+          {
+            contextItems: [],
+            isGatheringContext: false,
+            message: {
+              content: "I'll search the codebase for you.",
+              id: expect.any(String),
+              role: "assistant",
+              toolCalls: [
+                {
+                  id: "tool-call-1",
+                  type: "function",
+                  function: {
+                    name: "search_codebase",
+                    arguments: JSON.stringify({ query: "test function" }),
+                  },
+                },
+              ],
+            },
+            promptLogs: [
+              {
+                completion: "I'll search the codebase for you.",
+                modelProvider: "anthropic",
+                prompt: "Please search the codebase",
+              },
+            ],
+            toolCallStates: [
+              {
+                toolCallId: "tool-call-1",
+                toolCall: {
+                  id: "tool-call-1",
+                  type: "function",
+                  function: {
+                    name: "search_codebase",
+                    arguments: JSON.stringify({ query: "test function" }),
+                  },
+                },
+                parsedArgs: { query: "test function" },
+                status: "done",
+                output: [
+                  {
+                    name: "Search Results",
+                    description: "Found 3 matches",
+                    content: "Result 1\nResult 2\nResult 3",
+                    icon: "search",
+                    hidden: false,
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            contextItems: [],
+            message: {
+              content: "Result 1\nResult 2\nResult 3",
+              id: expect.any(String),
+              role: "tool",
+              toolCallId: "tool-call-1",
+            },
+          },
+          {
+            contextItems: [],
+            isGatheringContext: false,
+            message: {
+              content: "Search completed.",
+              id: expect.any(String),
+              role: "assistant",
+            },
+            promptLogs: [
+              {
+                completion: "Search completed.",
+                modelProvider: "anthropic",
+                prompt: "continuing after tool",
+              },
+            ],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: false,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController),
+        contextPercentage: 0.9,
+        isPruned: false,
+        isInEdit: false,
+        title: "New Session",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: {},
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {
+          search_codebase: "allowedWithoutPermission",
+        },
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
   });
+
+  it("should throw error when no chat model is selected", async () => {
+    const { mockStore, mockIdeMessenger } = setupTest();
+
+    // Create store with no selected chat model
+    const mockStoreNoModel = createMockStore({
+      session: {
+        history: [
+          {
+            message: { id: "1", role: "user", content: "Hello" },
+            contextItems: [],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: true,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: new AbortController(),
+        contextPercentage: 0,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: null, // No model selected
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        } as any,
+        lastSelectedModelByRole: {},
+      } as any,
+      ui: {
+        toolSettings: {},
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      } as any,
+    });
+
+    // Execute thunk and expect it to be rejected
+    const result = await mockStoreNoModel.dispatch(
+      streamNormalInput({}) as any,
+    );
+
+    // Verify the thunk was rejected with correct error
+    expect(result.type).toBe("chat/streamNormalInput/rejected");
+    expect(result.error?.message).toBe("No chat model selected");
+
+    // Verify only the pending action was dispatched (no other actions should occur)
+    const dispatchedActions = (mockStoreNoModel as any).getActions();
+    expect(dispatchedActions).toEqual([
+      {
+        type: "chat/streamNormalInput/pending",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "pending",
+        }),
+      },
+      {
+        type: "chat/streamNormalInput/rejected",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "rejected",
+        }),
+        error: expect.objectContaining({
+          message: "No chat model selected",
+        }),
+      },
+    ]);
+
+    // Verify no IDE messenger calls were made
+    expect(mockStoreNoModel.mockIdeMessenger.request).not.toHaveBeenCalled();
+    expect(
+      mockStoreNoModel.mockIdeMessenger.llmStreamChat,
+    ).not.toHaveBeenCalled();
+
+    // Verify state remains unchanged from initial mock store values
+    const finalState = mockStoreNoModel.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            message: { id: "1", role: "user", content: "Hello" },
+            contextItems: [],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: true, // Unchanged from initial
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController),
+        contextPercentage: 0, // Unchanged from initial
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined, // Unchanged from initial
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: null, // No model selected
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {},
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {},
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
+  });
+
+  it("should handle compilation error - not enough context", async () => {
+    const { mockStore, mockIdeMessenger } = setupTest();
+
+    // Mock compilation failure with "Not enough context" error
+    mockIdeMessenger.request.mockResolvedValue({
+      status: "error",
+      error: "Not enough context available for this request",
+    });
+
+    // Execute thunk
+    const result = await mockStore.dispatch(streamNormalInput({}) as any);
+
+    // Verify thunk completed successfully (doesn't throw, handles error gracefully)
+    expect(result.type).toBe("chat/streamNormalInput/fulfilled");
+
+    // Verify exact action sequence for this error path
+    const dispatchedActions = (mockStore as any).getActions();
+    expect(dispatchedActions).toEqual([
+      {
+        type: "chat/streamNormalInput/pending",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "pending",
+        }),
+      },
+      {
+        type: "session/setAppliedRulesAtIndex",
+        payload: {
+          index: 0,
+          appliedRules: [],
+        },
+      },
+      {
+        type: "session/setActive",
+      },
+      {
+        type: "session/setInlineErrorMessage",
+      },
+      // Error handling actions
+      {
+        type: "session/setInlineErrorMessage",
+        payload: "out-of-context",
+      },
+      {
+        type: "session/setInactive",
+      },
+      {
+        type: "chat/streamNormalInput/fulfilled",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "fulfilled",
+        }),
+      },
+    ]);
+
+    // Verify IDE messenger was called for compilation but not streaming
+    expect(mockIdeMessenger.request).toHaveBeenCalledWith("llm/compileChat", {
+      messages: [{ role: "user", content: "Hello" }],
+      options: {},
+    });
+    expect(mockIdeMessenger.llmStreamChat).not.toHaveBeenCalled();
+
+    // Verify final state shows error condition
+    const finalState = mockStore.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            appliedRules: [],
+            contextItems: [],
+            isGatheringContext: false,
+            message: { id: "1", role: "user", content: "Hello" },
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: false, // Should be inactive due to error
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController),
+        contextPercentage: 0, // Unchanged - no successful compilation
+        isPruned: false, // Unchanged - no successful compilation
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: "out-of-context", // Error message set
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {},
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
+  });
+
+  it("should throw error for other compilation errors", async () => {
+    const { mockStore, mockIdeMessenger } = setupTest();
+
+    // Mock compilation failure with generic error
+    mockIdeMessenger.request.mockResolvedValue({
+      status: "error",
+      error: "Model configuration is invalid",
+    });
+
+    // Execute thunk and expect it to be rejected
+    const result = await mockStore.dispatch(streamNormalInput({}) as any);
+
+    // Verify thunk was rejected with the compilation error
+    expect(result.type).toBe("chat/streamNormalInput/rejected");
+    expect(result.error?.message).toBe("Model configuration is invalid");
+
+    // Verify action sequence - should get to compilation, then fail
+    const dispatchedActions = (mockStore as any).getActions();
+    expect(dispatchedActions).toEqual([
+      {
+        type: "chat/streamNormalInput/pending",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "pending",
+        }),
+      },
+      {
+        type: "session/setAppliedRulesAtIndex",
+        payload: {
+          index: 0,
+          appliedRules: [],
+        },
+      },
+      {
+        type: "session/setActive",
+      },
+      {
+        type: "session/setInlineErrorMessage",
+      },
+      // Error causes thunk rejection - no cleanup actions
+      {
+        type: "chat/streamNormalInput/rejected",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "rejected",
+        }),
+        error: expect.objectContaining({
+          message: "Model configuration is invalid",
+        }),
+      },
+    ]);
+
+    // Verify IDE messenger was called for compilation but not streaming
+    expect(mockIdeMessenger.request).toHaveBeenCalledWith("llm/compileChat", {
+      messages: [{ role: "user", content: "Hello" }],
+      options: {},
+    });
+    expect(mockIdeMessenger.llmStreamChat).not.toHaveBeenCalled();
+
+    // Verify final state - error interrupted processing, so session remains active
+    const finalState = mockStore.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            appliedRules: [],
+            contextItems: [],
+            message: { id: "1", role: "user", content: "Hello" },
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: true, // TODO: fix this
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController),
+        contextPercentage: 0, // Unchanged - no successful compilation
+        isPruned: false, // Unchanged - no successful compilation
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined, // No special error message - error was thrown
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {},
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
+  });
+
+  it("should handle streaming abort", async () => {
+    // Create an AbortController that we'll abort during streaming
+    const testAbortController = new AbortController();
+
+    // Create store with our test abort controller
+    const mockStoreWithAbort = createMockStore({
+      session: {
+        history: [
+          {
+            message: { id: "1", role: "user", content: "Hello" },
+            contextItems: [],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: true,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: testAbortController, // Use our test controller
+        contextPercentage: 0,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        } as any,
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+      } as any,
+      ui: {
+        toolSettings: {},
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      } as any,
+    });
+
+    const mockIdeMessengerAbort = mockStoreWithAbort.mockIdeMessenger;
+
+    // Setup successful compilation
+    mockIdeMessengerAbort.request.mockResolvedValue({
+      status: "success",
+      content: {
+        compiledChatMessages: [{ role: "user", content: "Hello" }],
+        didPrune: false,
+        contextPercentage: 0.8,
+      },
+    });
+
+    // Setup streaming generator that simulates abort by checking isStreaming state
+    async function* mockStreamGeneratorWithAbort() {
+      yield [{ role: "assistant", content: "First chunk" }];
+
+      // Simulate user clicking abort button - this would dispatch setInactive
+      setTimeout(() => {
+        mockStoreWithAbort.dispatch({ type: "session/setInactive" });
+      }, 5);
+
+      // After first yield, wait a bit then yield second chunk (which should be ignored due to abort)
+      await new Promise((resolve) => setTimeout(resolve, 10));
+      yield [{ role: "assistant", content: "Second chunk" }];
+
+      return {
+        prompt: "Hello",
+        completion: "Complete response",
+        modelProvider: "anthropic",
+      };
+    }
+
+    mockIdeMessengerAbort.llmStreamChat.mockReturnValue(
+      mockStreamGeneratorWithAbort(),
+    );
+
+    // Execute thunk - should be aborted
+    const result = await mockStoreWithAbort.dispatch(
+      streamNormalInput({}) as any,
+    );
+
+    // Verify thunk completed successfully (abort just stops streaming early)
+    expect(result.type).toBe("chat/streamNormalInput/fulfilled");
+
+    // Verify exact action sequence - should start but then be aborted
+    const dispatchedActions = (mockStoreWithAbort as any).getActions();
+    expect(dispatchedActions).toEqual([
+      {
+        type: "chat/streamNormalInput/pending",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "pending",
+        }),
+      },
+      {
+        type: "session/setAppliedRulesAtIndex",
+        payload: {
+          index: 0,
+          appliedRules: [],
+        },
+      },
+      {
+        type: "session/setActive",
+      },
+      {
+        type: "session/setInlineErrorMessage",
+      },
+      {
+        type: "session/setIsPruned",
+        payload: false,
+      },
+      {
+        type: "session/setContextPercentage",
+        payload: 0.8,
+      },
+      {
+        type: "session/streamUpdate",
+        payload: [
+          {
+            role: "assistant",
+            content: "First chunk",
+          },
+        ],
+      },
+      // User abort action
+      {
+        type: "session/setInactive",
+      },
+      // Stream abort dispatch (called by implementation)
+      {
+        type: "session/abortStream",
+      },
+      // Stream end cleanup
+      {
+        type: "session/setInactive",
+      },
+      // Stream completes but no more updates are processed
+      {
+        type: "chat/streamNormalInput/fulfilled",
+        meta: expect.objectContaining({
+          arg: {},
+          requestStatus: "fulfilled",
+        }),
+      },
+    ]);
+
+    // Verify IDE messenger calls
+    expect(mockIdeMessengerAbort.request).toHaveBeenCalledWith(
+      "llm/compileChat",
+      {
+        messages: [{ role: "user", content: "Hello" }],
+        options: {},
+      },
+    );
+
+    expect(mockIdeMessengerAbort.llmStreamChat).toHaveBeenCalledWith(
+      expect.objectContaining({
+        completionOptions: {},
+        title: "Claude 3.5 Sonnet",
+        messages: [{ role: "user", content: "Hello" }],
+        messageOptions: { precompiled: true },
+      }),
+      expect.any(AbortSignal),
+    );
+
+    // Dev data logging should not occur since streaming was stopped early
+    expect(mockIdeMessengerAbort.post).not.toHaveBeenCalledWith(
+      "devdata/log",
+      expect.anything(),
+    );
+
+    // Verify final state - streaming should be stopped, partial content preserved
+    const finalState = mockStoreWithAbort.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            appliedRules: [],
+            contextItems: [],
+            message: { id: "1", role: "user", content: "Hello" },
+          },
+          {
+            contextItems: [],
+            isGatheringContext: false,
+            message: {
+              content: "First chunk", // Only first chunk before abort
+              id: expect.any(String),
+              role: "assistant",
+            },
+            // No promptLogs because streaming was stopped before completion
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: false, // False because setInactive was called
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController), // New controller after abort
+        contextPercentage: 0.8,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {},
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
+  });
+
+  it("should handle tool call rejection error", async () => {
+    const { mockStore, mockIdeMessenger } = setupTest();
+
+    // Create store with tool settings that will reject the tool call
+    const mockStoreWithToolReject = createMockStore({
+      session: {
+        history: [
+          {
+            message: {
+              id: "1",
+              role: "user",
+              content: "Please edit this file",
+            },
+            contextItems: [],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: true,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: new AbortController(),
+        contextPercentage: 0,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        } as any,
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+      } as any,
+      ui: {
+        toolSettings: {
+          edit_existing_file: "disabled", // Tool is disabled/rejected
+        },
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      } as any,
+    });
+
+    const mockIdeMessengerReject = mockStoreWithToolReject.mockIdeMessenger;
+
+    // Setup successful compilation
+    mockIdeMessengerReject.request.mockResolvedValue({
+      status: "success",
+      content: {
+        compiledChatMessages: [
+          { role: "user", content: "Please edit this file" },
+        ],
+        didPrune: false,
+        contextPercentage: 0.7,
+      },
+    });
+
+    // Setup streaming generator with tool call that will be rejected
+    async function* mockStreamGeneratorWithRejectedTool() {
+      yield [{ role: "assistant", content: "I'll edit the file for you." }];
+      yield [
+        {
+          role: "assistant",
+          content: "",
+          toolCalls: [
+            {
+              id: "tool-reject-1",
+              type: "function",
+              function: {
+                name: "edit_existing_file",
+                arguments: JSON.stringify({
+                  filepath: "test.js",
+                  changes: "const x = 1;",
+                }),
+              },
+            },
+          ],
+        },
+      ];
+      return {
+        prompt: "Please edit this file",
+        completion: "I'll edit the file for you.",
+        modelProvider: "anthropic",
+      };
+    }
+
+    mockIdeMessengerReject.llmStreamChat.mockReturnValue(
+      mockStreamGeneratorWithRejectedTool(),
+    );
+
+    // Execute thunk
+    const result = await mockStoreWithToolReject.dispatch(
+      streamNormalInput({}) as any,
+    );
+
+    // Verify thunk completed successfully (tool rejection is handled gracefully)
+    expect(result.type).toBe("chat/streamNormalInput/fulfilled");
+
+    // Verify action sequence includes tool generation but rejection handling
+    const dispatchedActions = (mockStoreWithToolReject as any).getActions();
+
+    // Should contain the standard streaming setup actions
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setAppliedRulesAtIndex",
+        payload: { index: 0, appliedRules: [] },
+      }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({ type: "session/setActive" }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setContextPercentage",
+        payload: 0.7,
+      }),
+    );
+
+    // Should contain streaming updates including the tool call
+    const streamUpdates = dispatchedActions.filter(
+      (action: any) => action.type === "session/streamUpdate",
+    );
+    expect(streamUpdates.length).toBeGreaterThanOrEqual(2);
+
+    // Tool call should be marked as generated but then rejected
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setToolGenerated",
+        payload: expect.objectContaining({ toolCallId: "tool-reject-1" }),
+      }),
+    );
+
+    // Should complete successfully despite tool rejection
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "chat/streamNormalInput/fulfilled",
+      }),
+    );
+
+    // Verify IDE messenger calls - compilation should succeed, streaming should happen
+    expect(mockIdeMessengerReject.request).toHaveBeenCalledWith(
+      "llm/compileChat",
+      {
+        messages: [{ role: "user", content: "Hello" }], // constructMessages mock
+        options: {},
+      },
+    );
+
+    expect(mockIdeMessengerReject.llmStreamChat).toHaveBeenCalled();
+
+    // Verify final state contains the tool call but it's not executed
+    const finalState = mockStoreWithToolReject.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            appliedRules: [],
+            contextItems: [],
+            message: {
+              id: "1",
+              role: "user",
+              content: "Please edit this file",
+            },
+          },
+          {
+            contextItems: [],
+            isGatheringContext: false,
+            message: {
+              content: "I'll edit the file for you.",
+              id: expect.any(String),
+              role: "assistant",
+              toolCalls: [
+                {
+                  id: "tool-reject-1",
+                  type: "function",
+                  function: {
+                    name: "edit_existing_file",
+                    arguments: JSON.stringify({
+                      filepath: "test.js",
+                      changes: "const x = 1;",
+                    }),
+                  },
+                },
+              ],
+            },
+            promptLogs: [
+              {
+                completion: "I'll edit the file for you.",
+                modelProvider: "anthropic",
+                prompt: "Please edit this file",
+              },
+            ],
+            toolCallStates: [
+              {
+                toolCallId: "tool-reject-1",
+                toolCall: {
+                  id: "tool-reject-1",
+                  type: "function",
+                  function: {
+                    name: "edit_existing_file",
+                    arguments: JSON.stringify({
+                      filepath: "test.js",
+                      changes: "const x = 1;",
+                    }),
+                  },
+                },
+                parsedArgs: {
+                  filepath: "test.js",
+                  changes: "const x = 1;",
+                },
+                status: "generated", // Tool call exists but isn't executed due to settings
+              },
+            ],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: false,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController),
+        contextPercentage: 0.7,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {
+          edit_existing_file: "disabled",
+        },
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
+  });
+
+  it("should handle tool call requiring manual approval", async () => {
+    const { mockStore, mockIdeMessenger } = setupTest();
+
+    // Create store with tool settings that require manual approval (ask first)
+    const mockStoreWithManualApproval = createMockStore({
+      session: {
+        history: [
+          {
+            message: {
+              id: "1",
+              role: "user",
+              content: "Please search the codebase",
+            },
+            contextItems: [],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: true,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: new AbortController(),
+        contextPercentage: 0,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,  
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        } as any,
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+      } as any,
+      ui: {
+        toolSettings: {
+          search_codebase: "askFirst", // Requires manual approval
+        },
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      } as any,
+    });
+
+    const mockIdeMessengerManual = mockStoreWithManualApproval.mockIdeMessenger;
+
+    // Setup successful compilation
+    mockIdeMessengerManual.request.mockResolvedValue({
+      status: "success",
+      content: {
+        compiledChatMessages: [
+          { role: "user", content: "Please search the codebase" },
+        ],
+        didPrune: false,
+        contextPercentage: 0.9,
+      },
+    });
+
+    // Setup streaming generator with tool call requiring approval
+    async function* mockStreamGeneratorWithApprovalTool() {
+      yield [
+        { role: "assistant", content: "I'll search the codebase for you." },
+      ];
+      yield [
+        {
+          role: "assistant",
+          content: "",
+          toolCalls: [
+            {
+              id: "tool-approval-1",
+              type: "function",
+              function: {
+                name: "search_codebase",
+                arguments: JSON.stringify({ query: "test function" }),
+              },
+            },
+          ],
+        },
+      ];
+      return {
+        prompt: "Please search the codebase",
+        completion: "I'll search the codebase for you.",
+        modelProvider: "anthropic",
+      };
+    }
+
+    mockIdeMessengerManual.llmStreamChat.mockReturnValue(
+      mockStreamGeneratorWithApprovalTool(),
+    );
+
+    // Execute thunk
+    const result = await mockStoreWithManualApproval.dispatch(
+      streamNormalInput({}) as any,
+    );
+
+    // Verify thunk completed successfully (tool waits for approval)
+    expect(result.type).toBe("chat/streamNormalInput/fulfilled");
+
+    // Verify action sequence includes tool generation but NO execution
+    const dispatchedActions = (mockStoreWithManualApproval as any).getActions();
+
+    // Should contain the standard streaming setup actions
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setAppliedRulesAtIndex",
+        payload: { index: 0, appliedRules: [] },
+      }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({ type: "session/setActive" }),
+    );
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setContextPercentage",
+        payload: 0.9,
+      }),
+    );
+
+    // Should contain streaming updates including the tool call
+    const streamUpdates = dispatchedActions.filter(
+      (action: any) => action.type === "session/streamUpdate",
+    );
+    expect(streamUpdates.length).toBeGreaterThanOrEqual(2);
+
+    // Tool call should be marked as generated but NOT executed (waiting for approval)
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setToolGenerated",
+        payload: expect.objectContaining({ toolCallId: "tool-approval-1" }),
+      }),
+    );
+
+    // Should NOT contain any tool execution actions
+    expect(dispatchedActions).not.toContainEqual(
+      expect.objectContaining({
+        type: "session/setToolCallCalling",
+      }),
+    );
+    expect(dispatchedActions).not.toContainEqual(
+      expect.objectContaining({
+        type: "session/updateToolCallOutput",
+      }),
+    );
+    expect(dispatchedActions).not.toContainEqual(
+      expect.objectContaining({
+        type: "session/acceptToolCall",
+      }),
+    );
+
+    // Should complete successfully
+    expect(dispatchedActions).toContainEqual(
+      expect.objectContaining({
+        type: "chat/streamNormalInput/fulfilled",
+      }),
+    );
+
+    // Verify IDE messenger calls - compilation should happen, streaming should happen
+    expect(mockIdeMessengerManual.request).toHaveBeenCalledWith(
+      "llm/compileChat",
+      {
+        messages: [{ role: "user", content: "Hello" }], // constructMessages mock
+        options: {},
+      },
+    );
+
+    expect(mockIdeMessengerManual.llmStreamChat).toHaveBeenCalled();
+
+    // Should NOT call tools/call since tool requires approval
+    expect(mockIdeMessengerManual.request).not.toHaveBeenCalledWith(
+      "tools/call",
+      expect.anything(),
+    );
+
+    // Verify final state contains the tool call in "generated" state (waiting for approval)
+    const finalState = mockStoreWithManualApproval.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            appliedRules: [],
+            contextItems: [],
+            message: {
+              id: "1",
+              role: "user",
+              content: "Please search the codebase",
+            },
+          },
+          {
+            contextItems: [],
+            isGatheringContext: false,
+            message: {
+              content: "I'll search the codebase for you.",
+              id: expect.any(String),
+              role: "assistant",
+              toolCalls: [
+                {
+                  id: "tool-approval-1",
+                  type: "function",
+                  function: {
+                    name: "search_codebase",
+                    arguments: JSON.stringify({ query: "test function" }),
+                  },
+                },
+              ],
+            },
+            promptLogs: [
+              {
+                completion: "I'll search the codebase for you.",
+                modelProvider: "anthropic",
+                prompt: "Please search the codebase",
+              },
+            ],
+            toolCallStates: [
+              {
+                toolCallId: "tool-approval-1",
+                toolCall: {
+                  id: "tool-approval-1",
+                  type: "function",
+                  function: {
+                    name: "search_codebase",
+                    arguments: JSON.stringify({ query: "test function" }),
+                  },
+                },
+                parsedArgs: { query: "test function" },
+                status: "generated", // Waiting for user approval
+              },
+            ],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: false,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController),
+        contextPercentage: 0.9,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {
+          search_codebase: "askFirst",
+        },
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
+  });
+
+  it("should handle complete user approval and tool execution flow", async () => {
+    const { mockStore, mockIdeMessenger } = setupTest();
+
+    // Create store with tool settings that require manual approval
+    const mockStoreWithApproval = createMockStore({
+      session: {
+        history: [
+          {
+            message: {
+              id: "1",
+              role: "user",
+              content: "Please search the codebase for test functions",
+            },
+            contextItems: [],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: true,
+        id: "session-123",
+        mode: "chat",
+        streamAborter: new AbortController(),
+        contextPercentage: 0,
+        isPruned: false,
+        isInEdit: false,
+        title: "",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: [],
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        } as any,
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+      } as any,
+      ui: {
+        toolSettings: {
+          search_codebase: "askFirst", // Requires manual approval
+        },
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      } as any,
+    });
+
+    const mockIdeMessengerApproval = mockStoreWithApproval.mockIdeMessenger;
+
+    // Setup successful compilation
+    mockIdeMessengerApproval.request.mockImplementation((endpoint: string, data: any) => {
+      if (endpoint === "llm/compileChat") {
+        return Promise.resolve({
+          status: "success",
+          content: {
+            compiledChatMessages: [
+              { role: "user", content: "Please search the codebase for test functions" },
+            ],
+            didPrune: false,
+            contextPercentage: 0.85,
+          },
+        });
+      } else if (endpoint === "tools/call") {
+        // Mock server-side tool call response for search_codebase
+        return Promise.resolve({
+          status: "success",
+          content: {
+            contextItems: [
+              {
+                name: "Search Results",
+                description: "Found test functions",
+                content: "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
+                icon: "search",
+                hidden: false,
+              },
+            ],
+            errorMessage: undefined,
+          },
+        });
+      } else if (endpoint === "history/save") {
+        return Promise.resolve({ status: "success" });
+      }
+      return Promise.resolve({ status: "success", content: {} });
+    });
+
+    // Setup streaming generator with tool call
+    async function* mockStreamGeneratorWithApprovalFlow() {
+      yield [
+        { role: "assistant", content: "I'll search for test functions in the codebase." },
+      ];
+      yield [
+        {
+          role: "assistant",
+          content: "",
+          toolCalls: [
+            {
+              id: "tool-approval-flow-1",
+              type: "function",
+              function: {
+                name: "search_codebase",
+                arguments: JSON.stringify({ query: "test function" }),
+              },
+            },
+          ],
+        },
+      ];
+      return {
+        prompt: "Please search the codebase for test functions",
+        completion: "I'll search for test functions in the codebase.",
+        modelProvider: "anthropic",
+      };
+    }
+
+    // Mock subsequent streaming calls (after tool execution)
+    let streamCallCount = 0;
+    mockIdeMessengerApproval.llmStreamChat.mockImplementation(() => {
+      streamCallCount++;
+      if (streamCallCount === 1) {
+        // First call - initial streaming with tool call
+        return mockStreamGeneratorWithApprovalFlow();
+      } else {
+        // Subsequent calls from streamResponseAfterToolCall
+        async function* followupGenerator() {
+          yield [
+            { 
+              role: "assistant", 
+              content: "I found several test functions in your codebase. Here are the main ones I discovered..." 
+            },
+          ];
+          return {
+            prompt: "continuing after tool execution",
+            completion: "I found several test functions in your codebase. Here are the main ones I discovered...",
+            modelProvider: "anthropic",
+          };
+        }
+        return followupGenerator();
+      }
+    });
+
+    // Execute initial thunk - this should generate the tool call but not execute it
+    const initialResult = await mockStoreWithApproval.dispatch(
+      streamNormalInput({}) as any,
+    );
+
+    // Verify initial streaming completed successfully 
+    expect(initialResult.type).toBe("chat/streamNormalInput/fulfilled");
+
+    // Get initial actions to verify tool was generated
+    const initialActions = (mockStoreWithApproval as any).getActions();
+    expect(initialActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setToolGenerated",
+        payload: expect.objectContaining({ toolCallId: "tool-approval-flow-1" }),
+      }),
+    );
+
+    // Clear the actions array to track only the approval flow
+    (mockStoreWithApproval as any).clearActions();
+
+    // Import the callToolById thunk to simulate user approval
+    const { callToolById } = await import("./callToolById");
+
+    // Simulate user clicking "Accept" on the tool call
+    const approvalResult = await mockStoreWithApproval.dispatch(
+      callToolById({ toolCallId: "tool-approval-flow-1" }) as any,
+    );
+
+    // Verify tool execution completed successfully
+    expect(approvalResult.type).toBe("chat/callTool/fulfilled");
+
+    // Verify the approval flow actions
+    const approvalActions = (mockStoreWithApproval as any).getActions();
+
+    // Should contain tool execution actions
+    expect(approvalActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/setToolCallCalling",
+        payload: { toolCallId: "tool-approval-flow-1" },
+      }),
+    );
+
+    expect(approvalActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/updateToolCallOutput",
+        payload: {
+          toolCallId: "tool-approval-flow-1",
+          contextItems: [
+            {
+              name: "Search Results",
+              description: "Found test functions",
+              content: "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
+              icon: "search",
+              hidden: false,
+            },
+          ],
+        },
+      }),
+    );
+
+    expect(approvalActions).toContainEqual(
+      expect.objectContaining({
+        type: "session/acceptToolCall",
+        payload: { toolCallId: "tool-approval-flow-1" },
+      }),
+    );
+
+    // Should trigger follow-up streaming (streamAfterToolCall)
+    expect(approvalActions).toContainEqual(
+      expect.objectContaining({
+        type: "chat/streamAfterToolCall/fulfilled",
+      }),
+    );
+
+    // Verify IDE messenger calls for tool execution
+    expect(mockIdeMessengerApproval.request).toHaveBeenCalledWith("tools/call", {
+      toolCall: {
+        id: "tool-approval-flow-1",
+        type: "function",
+        function: {
+          name: "search_codebase",
+          arguments: JSON.stringify({ query: "test function" }),
+        },
+      },
+    });
+
+    // Verify second streaming call was made for continuation
+    expect(streamCallCount).toBe(2);
+
+    // Verify final state shows completed tool call and follow-up response
+    const finalState = mockStoreWithApproval.getState();
+    expect(finalState).toEqual({
+      session: {
+        history: [
+          {
+            appliedRules: [],
+            contextItems: [],
+            message: {
+              id: "1",
+              role: "user",
+              content: "Please search the codebase for test functions",
+            },
+          },
+          {
+            contextItems: [],
+            isGatheringContext: false,
+            message: {
+              content: "I'll search for test functions in the codebase.",
+              id: expect.any(String),
+              role: "assistant",
+              toolCalls: [
+                {
+                  id: "tool-approval-flow-1",
+                  type: "function",
+                  function: {
+                    name: "search_codebase",
+                    arguments: JSON.stringify({ query: "test function" }),
+                  },
+                },
+              ],
+            },
+            promptLogs: [
+              {
+                completion: "I'll search for test functions in the codebase.",
+                modelProvider: "anthropic",
+                prompt: "Please search the codebase for test functions",
+              },
+            ],
+            toolCallStates: [
+              {
+                toolCallId: "tool-approval-flow-1",
+                toolCall: {
+                  id: "tool-approval-flow-1",
+                  type: "function",
+                  function: {
+                    name: "search_codebase",
+                    arguments: JSON.stringify({ query: "test function" }),
+                  },
+                },
+                parsedArgs: { query: "test function" },
+                status: "done", // Tool call completed successfully
+                output: [
+                  {
+                    name: "Search Results",
+                    description: "Found test functions",
+                    content: "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
+                    icon: "search",
+                    hidden: false,
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            contextItems: [],
+            message: {
+              content: "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
+              id: expect.any(String),
+              role: "tool",
+              toolCallId: "tool-approval-flow-1",
+            },
+          },
+          {
+            contextItems: [],
+            isGatheringContext: false,
+            message: {
+              content: "I found several test functions in your codebase. Here are the main ones I discovered...",
+              id: expect.any(String),
+              role: "assistant",
+            },
+            promptLogs: [
+              {
+                completion: "I found several test functions in your codebase. Here are the main ones I discovered...",
+                modelProvider: "anthropic",
+                prompt: "continuing after tool execution",
+              },
+            ],
+          },
+        ],
+        hasReasoningEnabled: false,
+        isStreaming: false, // Inactive after complete flow
+        id: "session-123",
+        mode: "chat",
+        streamAborter: expect.any(AbortController),
+        contextPercentage: 0.85,
+        isPruned: false,
+        isInEdit: false,
+        title: "New Session",
+        lastSessionId: undefined,
+        isSessionMetadataLoading: false,
+        allSessionMetadata: {},
+        symbols: {},
+        codeBlockApplyStates: {
+          states: [],
+          curIndex: 0,
+        },
+        newestToolbarPreviewForInput: {},
+        compactionLoading: {},
+        inlineErrorMessage: undefined,
+      },
+      config: {
+        config: {
+          tools: [],
+          rules: [],
+          tabAutocompleteModel: undefined,
+          selectedModelByRole: {
+            chat: mockClaudeModel,
+            apply: null,
+            edit: null,
+            summarize: null,
+            autocomplete: null,
+            rerank: null,
+            embed: null,
+          },
+          experimental: {
+            onlyUseSystemMessageTools: false,
+          },
+        },
+        lastSelectedModelByRole: {
+          chat: mockClaudeModel.title,
+        },
+        loading: false,
+        configError: undefined,
+      },
+      ui: {
+        toolSettings: {
+          search_codebase: "askFirst",
+        },
+        ruleSettings: {},
+        showDialog: false,
+        dialogMessage: undefined,
+      },
+      editModeState: {
+        isInEdit: false,
+        returnToMode: "chat",
+      },
+      indexing: {
+        indexingState: "disabled",
+      },
+      tabs: {
+        tabsItems: [],
+      },
+      profiles: {
+        profiles: [],
+      },
+    });
+  });
+
 });
